refactor all examples into subdirectories.  instructions says to put underactuated/src in your PYTHONPATH
