+++ .circleci/config.yml
@@ -14,7 +14,7 @@ jobs:
           name: setup
           command: ./scripts/continuous_integration/circle_ci/setup
 
-      - run: echo 'export PYTHONPATH=/opt/drake/lib/python2.7/site-packages:${PYTHONPATH}' >> $BASH_ENV
+      - run: echo 'export PYTHONPATH=/root/underactuated/src:/opt/drake/lib/python2.7/site-packages:${PYTHONPATH}' >> $BASH_ENV
 
       - run:
           name: build and test
@@ -36,7 +36,7 @@ jobs:
           name: setup
           command: ./scripts/continuous_integration/circle_ci/setup
 
-      - run: echo 'export PYTHONPATH=/opt/drake/lib/python2.7/site-packages:${PYTHONPATH}' >> $BASH_ENV
+      - run: echo 'export PYTHONPATH=/root/underactuated/src:/opt/drake/lib/python2.7/site-packages:${PYTHONPATH}' >> $BASH_ENV
 
       - run:
           name: build and test
