+++ drake_version.json
@@ -1,7 +1,7 @@
 {
   "base_url" : "https://drake-packages.csail.mit.edu/drake/",
-  "build" : "nightly",
-  "version" : "20180217",
+  "build" : "continuous",
+  "version" : "latest",
   "__comment__build_version__options__" : {
     "nightly" : "'latest' or a date, e.g. '20170719'.  Nightlies have the date of the prior evening (not the morning after)",
     "continuous" : "'latest'",
