+++ src/double_integrator/CMakeLists.txt
@@ -1,0 +1,2 @@
+
+add_pytest(value_iteration.py)
