+++ src/double_pendulum/CMakeLists.txt
@@ -1,0 +1,9 @@
+
+# Note: no duration arguments to keep this first textbook example simple.
+# Just pay the 10second cost.
+add_pytest(simulator.py)
+
+add_nbtest(simulator.ipynb)
+
+add_pytest(manipulator.py)
+add_pytest(as_single_pendulum.py --duration 1.0)
