+++ src/double_pendulum/simulator.py
@@ -3,7 +3,7 @@ from pydrake.all import (BasicVector, DiagramBuilder, FloatingBaseType,
 from underactuated import (FindResource, PlanarRigidBodyVisualizer)
 
 # Load the double pendulum from Universal Robot Description Format
-tree = RigidBodyTree(FindResource("double_pendulum.urdf"),
+tree = RigidBodyTree(FindResource("double_pendulum/double_pendulum.urdf"),
                      FloatingBaseType.kFixed)
 
 # Set up a block diagram with the robot (dynamics) and a visualization block.
