+++ src/pendulum/CMakeLists.txt
@@ -1,0 +1,4 @@
+
+add_pytest(torque_slider_demo.py --duration 1.0)
+
+add_pytest(value_iteration.py)
