+++ src/pendulum/torque_slider_demo.py
