+++ src/simple/continuous_time_system.cc
