+++ src/simple/continuous_time_system_simulation.py
@@ -1,6 +1,6 @@
 import matplotlib.pyplot as plt
 from pydrake.all import (DiagramBuilder, SignalLogger, Simulator)
-from simple_continuous_time_system import *
+from continuous_time_system import *
 
 # Create a simple block diagram containing our system.
 builder = DiagramBuilder()
